create_flaunch_token.py
#!/usr/bin/env python3
"""
Flaunch Token Creation Script
Create your own token using the Flaunch protocol on Base
"""

import os
import json
from datetime import datetime
from web3 import Web3
from dotenv import load_dotenv

def load_env_vars():
    """Load environment variables"""
    load_dotenv()

    return {
        'private_key': os.getenv('ETHEREUM_PRIVATE_KEY'),
        'rpc_url': 'https://mainnet.base.org'
    }

def get_flaunch_contracts():
    """Get Flaunch contract addresses on Base mainnet"""
    # From Flaunch documentation
    contracts = {
        'FlaunchZap': '0x...',  # Need to find this
        'TreasuryManagerFactory': '0x48af8b28DDC5e5A86c4906212fc35Fa808CA8763',
        'RevenueManager': '0x1216c723853dac0449c01d01d6e529d751d9c0c8'
    }

    return contracts

def create_flaunch_token_params(name, symbol, description=""):
    """Create parameters for flaunching a token"""

    # Basic token parameters
    token_params = {
        'name': name,
        'symbol': symbol,
        'description': description or f"{name} - Created with Flaunch protocol",
        'image': f"https://via.placeholder.com/400x400?text={symbol}",
        'external_url': "https://flaunch.gg"
    }

    return token_params

def estimate_flaunch_cost():
    """Estimate the cost to flaunch a token"""
    # Base on Ethereum gas costs
    estimates = {
        'gas_limit': 500000,
        'gas_price_gwei': 0.1,
        'eth_cost': 0.00005,  # Approximate
        'usd_cost': 0.15      # Approximate at $3000/ETH
    }

    return estimates

def main():
    print('🚀 FLAUNCH TOKEN CREATION')
    print('=' * 40)
    print(f'Timestamp: {datetime.now().strftime("%Y-%m-%d %H:%M:%S")}')
    print()

    # Load environment
    env_vars = load_env_vars()

    if not env_vars['private_key']:
        print("❌ Missing ETHEREUM_PRIVATE_KEY in environment")
        print("Please set your private key in the .env file")
        return

    print('✅ Environment configured')
    print()

    # Get contract addresses
    contracts = get_flaunch_contracts()
    print('📋 FLAUNCH CONTRACTS:')
    for name, address in contracts.items():
        print(f'  {name}: {address}')
    print()

    # Cost estimation
    costs = estimate_flaunch_cost()
    print('💰 ESTIMATED COSTS:')
    print(f'  Gas Limit: {costs["gas_limit"]:,}')
    print(f'  Gas Price: {costs["gas_price_gwei"]} gwei')
    print(f'  ETH Cost: {costs["eth_cost"]:.6f} ETH')
    print(f'  USD Cost: ${costs["usd_cost"]:.2f} USD')
    print()

    print('🎯 FLAUNCH TOKEN CREATION PROCESS:')
    print()
    print('1. PREPARATION:')
    print('   • Choose token name and symbol')
    print('   • Prepare token metadata (image, description)')
    print('   • Ensure wallet has sufficient ETH for gas')
    print('   • Set up revenue sharing (optional)')
    print()

    print('2. TOKEN MANAGER SETUP:')
    print('   • Deploy RevenueManager (or custom manager)')
    print('   • Configure revenue sharing parameters')
    print('   • Set protocol fee (0-100%)')
    print()

    print('3. TOKEN LAUNCH:')
    print('   • Call FlaunchZap.flaunch()')
    print('   • Token created with automatic liquidity')
    print('   • Revenue management activated')
    print()

    print('4. POST-LAUNCH:')
    print('   • Monitor trading activity')
    print('   • Collect revenue from fees')
    print('   • Engage community')
    print()

    print('📊 REVENUE MODEL OPTIONS:')
    print()
    print('REVENUE MANAGER:')
    print('• Collects 100% of trading fees')
    print('• Configurable protocol fee (0-10%)')
    print('• Automatic ETH distribution')
    print('• Multiple token support')
    print()

    print('CUSTOM MANAGER:')
    print('• Full control over revenue logic')
    print('• Custom fee structures')
    print('• Integration with other protocols')
    print('• Advanced automation')
    print()

    print('🔧 TECHNICAL REQUIREMENTS:')
    print()
    print('SMART CONTRACTS:')
    print('• FlaunchZap: Main launch contract')
    print('• Token Manager: Revenue collection')
    print('• ERC20 Token: Standard token contract')
    print('• Uniswap V4 Pool: Automatic liquidity')
    print()

    print('PARAMETERS NEEDED:')
    print('• Token name and symbol')
    print('• Initial supply distribution')
    print('• Fee configuration')
    print('• Manager implementation')
    print()

    # Example token creation
    print('💡 EXAMPLE: CREATE YOUR FIRST FLAUNCH TOKEN')
    print()

    example_token = create_flaunch_token_params(
        name="My Awesome Token",
        symbol="AWESOME",
        description="The most awesome token on Base, created with Flaunch!"
    )

    print('Token Parameters:')
    for key, value in example_token.items():
        print(f'  {key}: {value}')
    print()

    print('⚠️  IMPORTANT NOTES:')
    print('• Flaunch tokens are experimental')
    print('• DYOR before launching')
    print('• Test on Base Sepolia first')
    print('• Understand revenue implications')
    print('• Comply with local regulations')
    print()

    print('🛠️  DEVELOPMENT RESOURCES:')
    print('• Flaunch SDK: https://www.npmjs.com/package/@flaunch/sdk')
    print('• Builders Dashboard: https://builders.flaunch.gg')
    print('• Documentation: https://docs.flaunch.gg')
    print('• Discord Support: https://discord.com/invite/flaunch')
    print()

    print('🚀 READY TO CREATE YOUR FLAUNCH TOKEN?')
    print()
    print('Next steps:')
    print('1. Set up your development environment')
    print('2. Get test ETH on Base Sepolia')
    print('3. Deploy a test token')
    print('4. Launch on mainnet when ready')
    print()

if __name__ == '__main__':
    main()
